Pull out HistogramImpl and HistogramStatisticsImpl into their own file.
