+++ source/common/stats/heap_stat_data.h
@@ -4,8 +4,10 @@
 #include <string>
 #include <unordered_set>
 
+#include "common/common/hash.h"
+#include "common/common/thread.h"
+#include "common/common/thread_annotations.h"
 #include "common/stats/stat_data_allocator.h"
-#include "common/stats/stats_impl.h"
 
 namespace Envoy {
 namespace Stats {
