+++ source/common/stats/isolated_store_impl.h
@@ -8,6 +8,7 @@
 #include "common/common/utility.h"
 #include "common/stats/heap_stat_data.h"
 #include "common/stats/stats_impl.h"
+#include "common/stats/stats_options_impl.h"
 #include "common/stats/utility.h"
 
 namespace Envoy {
