+++ source/common/stats/thread_local_store.h
@@ -11,8 +11,11 @@
 #include "envoy/thread_local/thread_local.h"
 
 #include "common/stats/heap_stat_data.h"
+#include "common/stats/histogram.h"
 #include "common/stats/stats_impl.h"
 #include "common/stats/utility.h"
+
+#include "circllhist.h"
 
 namespace Envoy {
 namespace Stats {
