+++ source/server/options_impl.h
@@ -7,7 +7,7 @@
 #include "envoy/common/exception.h"
 #include "envoy/server/options.h"
 
-#include "common/stats/stats_impl.h"
+#include "common/stats/stats_options_impl.h"
 
 #include "spdlog/spdlog.h"
