+++ test/extensions/stats_sinks/metrics_service/BUILD
@@ -36,6 +36,7 @@ envoy_extension_cc_test(
         "//source/common/buffer:zero_copy_input_stream_lib",
         "//source/common/grpc:codec_lib",
         "//source/common/grpc:common_lib",
+        "//source/common/stats:histogram_lib",
         "//source/extensions/stat_sinks/metrics_service:config",
         "//test/common/grpc:grpc_client_integration_lib",
         "//test/integration:http_integration_lib",
