+++ test/integration/websocket_integration_test.cc
@@ -13,6 +13,7 @@
 #include "test/test_common/printers.h"
 #include "test/test_common/utility.h"
 
+#include "absl/strings/str_cat.h"
 #include "gtest/gtest.h"
 
 using testing::MatchesRegex;
