+++ test/mocks/stats/mocks.h
@@ -10,6 +10,7 @@
 #include "envoy/thread_local/thread_local.h"
 #include "envoy/upstream/cluster_manager.h"
 
+#include "common/stats/histogram.h"
 #include "common/stats/isolated_store_impl.h"
 #include "common/stats/stats_impl.h"
