+++ test/test_common/utility.h
@@ -14,6 +14,7 @@
 
 #include "common/buffer/buffer_impl.h"
 #include "common/common/c_smart_ptr.h"
+#include "common/common/thread.h"
 #include "common/http/header_map_impl.h"
 #include "common/protobuf/utility.h"
 #include "common/stats/raw_stat_data.h"
